rewrite(web): slice 9r.2 round-1 review fixes
Multi-agent review of PR #671 surfaced two must-fix items, one
medium-severity security note, and two architecture nudges worth
doing now (cheap) rather than after 9r.3 lands more code on top.

What changed
------------
- **Race fix in <Login/> click handler.** The previous handler did
  `set_phase.set(Idle)` synchronously and relied on the action's
  async closure to follow up with `set_phase.set(InFlight)`. Between
  those two writes the button was briefly enabled, long enough for
  a synthetic double-click to dispatch a second concurrent
  ceremony. Now the handler sets `InFlight` directly and the
  redundant write inside the action future is removed — the
  disabled re-render is synchronous with the click event.
  Correctness reviewer's must-fix.

- **Dedupe Header status closures.** `status_attr` and `status_label`
  in `<Header/>` were two identical closures producing
  `"connected"`/`"disconnected"`. Folded into one closure used for
  both the `data-status` attribute and the visible label — one
  reactive subscription, no risk of silent divergence if one ever
  changed. Code-quality reviewer's must-fix.

- **Harden `js_err()` against debug-format leaks.** The previous
  fallback was `format!("{v:?}")` which can surface internal
  browser/extension state into the user-visible error string.
  Replaced with a generic `"unexpected browser error"` message;
  the debug form goes to `console.warn` for operator triage. This
  was the security reviewer's MEDIUM finding — bounded impact (only
  the signed-in user's screen) but unnecessary information in the
  DOM.

- **Drop unused `serde_json` dep.** The crate doesn't call
  `serde_json::` directly; `gloo-net`'s `.json(...)` pulls it in
  transitively. Listing it was misleading. Vision + code-quality
  reviewers both flagged.

- **Extract `<Login/>` to `crates/web/src/login.rs`.** `main.rs`
  was 454 lines holding shell + login + ceremony + state machine
  + helpers. Login + signin + LoginPhase + js_err + url_setup_token
  are cohesive enough to be their own module, and doing it now
  while the module is small is cheaper than after 9r.3 (pair
  ceremony) and 9r.4 (session-restore) land more code on top.
  Architecture reviewer's recommendation.

- **Extract auth e2e to `test/e2e-rust/auth.e2e.js`.** Same logic:
  layout tests and ceremony tests are different concerns; with
  9r.3 + 9r.4 about to add more, splitting now keeps each file
  focused. `smoke.e2e.js` shrinks back to layout + bundle plumbing.
  Architecture reviewer's recommendation.

- **Rename `sign-in surfaces server error on fresh install` →
  `sign-in renders error region on server rejection`.** The old
  title implied the test relied on server state; it actually
  stubs the route via `page.route()`. Code-quality reviewer's nit.

Deferred (not in this round)
----------------------------
- `signin()` split into `fetch_challenge()` + `post_assertion()` —
  nice-to-have, defer until 9r.3 needs to share the HTTP glue
  with the pair ceremony (which it will). Splitting now would be
  premature shape — the actual reuse pattern only emerges with a
  second consumer.
- `webauthn-rs-proto` exact-version pin — broader supply-chain
  audit, not a 9r.2 concern.
- `setup_token` format validation — 9r.3 territory; the value
  isn't sent anywhere in 9r.2.
- `katulong-wire` shared crate — architecture reviewer explicitly
  said "premature until 9r.3 + 9r.4 stabilize the wire shape."
- Pair-mode UI scaffolding — the device-name input was landed in
  9r.1 as a deliberate visual checkpoint; ripping it out and
  re-landing it in 9r.3 would be churn.

Verified
--------
- `cargo check --target wasm32-unknown-unknown` — clean
- `trunk build --release` — 459 KB WASM (unchanged)
- 7/7 e2e tests pass (split: 3 in smoke.e2e.js, 4 in auth.e2e.js)
